Adding All Part I Projects
Project 2 builds a private blockchain. In that project it is created
what could be the base of any blockchain platform using levelDB to store
the block data.
Project 3 builds a REST API for this private blockchain that allow users
to interact with blockchain application.
Project 4 codes an integral part into this blockchain - the Mempool
component. Additionally, demonstrates how to store any data in the
blockchain by Encoding and Decoding it.
Project 5 builds a DApp over Etherium (Star Notary) allowing one to
register a star claimming it.